Reflector working in most tests
