+++ lib/src/reflector.dart
@@ -1,0 +1,11 @@
+import 'package:reflectable/reflectable.dart';
+
+class Reflector extends Reflectable {
+  const Reflector()
+      : super(
+          invokingCapability,
+          declarationsCapability,
+          instanceInvokeCapability,
+          metadataCapability,
+        );
+}
