+++ pubspec.yaml
@@ -13,6 +13,7 @@ dev_dependencies:
   test: ^1.6.0
   reflectable: ^2.2.0
   build_runner: ^1.7.2
+  build_test: ^0.10.11
 
 dependencies:
   path: ^1.6.4
